Fix mixin name validation not running. (#4576)
* Fix mixin name validation not running.

* Fix caching of validateMixinNames task.

(cherry picked from commit 3e6c1f7d52d0ad63158880259115fc4b7c249de7)
