+++ fabric-client-gametest-api-v1/src/client/resources/fabric-client-gametest-api-v1.mixins.json
@@ -4,7 +4,7 @@
   "compatibilityLevel": "JAVA_21",
   "mixins": [
     "ClientChunkManagerAccessor",
-    "ClientChunkMapAccessor",
+    "ClientChunkManagerClientChunkMapAccessor",
     "ClientWorldAccessor",
     "CyclingButtonWidgetAccessor",
     "ScreenAccessor",
@@ -25,7 +25,7 @@
     "threading.ClientConnectionMixin",
     "threading.MinecraftClientMixin",
     "threading.MinecraftServerMixin",
-    "threading.ServerMainMixin",
+    "threading.MainMixin",
     "threading.ThreadExecutorMixin",
     "world.CreateWorldScreenAccessor",
     "world.CreateWorldScreenMixin"
