+++ fabric-content-registries-v0/src/main/resources/fabric-content-registries-v0.mixins.json
@@ -3,8 +3,8 @@
   "package": "net.fabricmc.fabric.mixin.content.registry",
   "compatibilityLevel": "JAVA_21",
   "mixins": [
-    "AbstractBlockMixin",
-    "AbstractBlockStateMixin",
+    "AbstractBlockAccessor",
+    "AbstractBlockAbstractBlockStateMixin",
     "AxeItemAccessor",
     "BrewingRecipeRegistryBuilderMixin",
     "FarmerWorkTaskAccessor",
