+++ fabric-data-attachment-api-v1/src/main/resources/fabric-data-attachment-api-v1.mixins.json
@@ -9,7 +9,7 @@
     "ChunkDataSenderMixin",
     "ChunkMixin",
     "ClientConnectionMixin",
-    "CustomPayloadS2CPacketAccessor",
+    "CustomPayloadC2SPacketAccessor",
     "EntityMixin",
     "SerializedChunkMixin",
     "ServerWorldMixin",
