+++ fabric-data-generation-api-v1/src/client/resources/fabric-data-generation-api-v1.client.mixins.json
@@ -4,7 +4,7 @@
   "compatibilityLevel": "JAVA_21",
   "client": [
     "MinecraftClientMixin",
-    "ModelProviderBlockDefinitionsMixin",
+    "ModelProviderBlockStateSuppliersMixin",
     "ModelProviderItemAssetsMixin",
     "ModelProviderMixin"
   ],
