+++ fabric-renderer-indigo/src/client/resources/fabric-renderer-indigo.mixins.json
@@ -10,7 +10,7 @@
     "ChunkRendererRegionMixin",
     "ItemRendererAccessor",
     "ItemRenderStateMixin",
-    "LayerRenderStateMixin",
+    "ItemRenderStateLayerRenderStateMixin",
     "SectionBuilderMixin"
   ],
   "injectors": {
